missions/week 5/back/ABLY-externship-5/requirements/base.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
boto3==1.20.41
botocore==1.23.41
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-bootstrap-v5==1.0.8
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-mathfilters==1.0.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.3
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
